requirements.txt
cachetools
einops
ipdb
nltk
omegaconf
packaging
pytorch_lightning==2.2.2
PyYAML
tqdm
SentencePiece
transformers==4.40.0
deepspeed==0.14.1
torch==2.2.2
peft==0.10.0
numpy
tensorboard
accelerate==0.29.3
dvc
wandb
bitsandbytes
torchmetrics
matplotlib
seaborn
scikit-learn